Split out OAC Fetcher
